started refactoring for SSR
