+++ components/styles/module/Page.module.css
@@ -1,8 +1,6 @@
 .pageContainer {
     background-color: var(--bg-primary);
     color: var(--text-primary);
-    min-height: 100vh;
-    min-width: 100vw;
 }
 
 .menuToggleButton {
