+++ lib/index.js
@@ -1,0 +1,8 @@
+// Pure utilities (safe for server & client)
+export { scrollToSection } from './navigation';
+export { copyToClipboard } from './clipboard';
+export { getMemberId, getEpId } from './ids';
+
+// Client-only hooks
+export { useOrientation } from './hooks/useOrientation';
+export { useMounted } from './hooks/useMounted';
